fix: implement bot utils
