--- constants/assets/colors/colors.ts
@@ -1,12 +1,0 @@
-import { ColorResolvable } from "discord.js";
-
-const colors = {
-  error: "#e44c3c" as ColorResolvable,
-  succesfull: "#37E001" as ColorResolvable,
-  default: "#4647EB" as ColorResolvable,
-  partners: {
-    serverblaze: "" as ColorResolvable,
-  },
-};
-
-export default colors;
