--- constants/assets/img/images.ts
@@ -1,29 +1,0 @@
-const images = {
-  avatar:
-    "https://cdn.discordapp.com/avatars/748215564455116961/ff37be1ab3cdf46c6c4179dcc9c11a91.png?size=1024",
-  logo: {
-    png: {
-      url: "https://cdn.discordapp.com/attachments/923519972288897024/923520268893319229/logo.png",
-    },
-    url: "https://cdn.discordapp.com/attachments/923519972288897024/923520268893319229/logo.png",
-  },
-  sponsors: {
-    serverblaze: {
-      logo_light: "https://serverblaze.eu/assets/mainpage/images/logo-light.png",
-    },
-    uptimerobot: {
-      banners: {
-        white:
-          "https://cdn.discordapp.com/attachments/844189583959916614/844192197733908510/UptimeRobotBannerWhite.png",
-        black:
-          "https://cdn.discordapp.com/attachments/844189583959916614/844192193593475142/UptimeRobotBannerBlack.png",
-      },
-      logo: {
-        white:
-          "https://cdn.discordapp.com/attachments/844189583959916614/844192200178139136/UptimeRobotShort.png",
-      },
-    },
-  },
-};
-
-export default images;
