--- constants/constants.ts
@@ -1,9 +1,0 @@
-import assets from "./assets/assets.js";
-import emojis from "./emojis/emojis.js";
-
-const constants = {
-  assets,
-  emojis,
-};
-
-export default constants;
