--- constants/emojis/emojis.ts
@@ -1,34 +1,0 @@
-const emojis = {
-  error: "<a:EazyBotsError:825828334164377653>",
-  succes: "<a:EazyBotsYes:884461991832465468>",
-  yes: "<a:EazyBotsYes:884461991832465468>",
-  statusOnline: "<a:EazyBotsStatusOnline:837715308240306197>",
-  loading: "<a:EazyBotsLoading:825818707455901719>",
-  ram: "<:EazyBotsRAM:825817433415483393>",
-  root: "<:EazyBotsRoot:825818176435912724>",
-  host: "<:EazyBotsRoot:825818176435912724>",
-  blank: "<:bla:84462458963756564>",
-  nodeJS: "<:EazyBotsNodeJS:838388312380932116>",
-  dev: "<:EazyBotsDeveloper:838389685289484289>",
-  err: "<a:EazyBotsError:825828334164377653>",
-  uptimerobot: "<:uptimeRobot:883984935407456266>",
-  calender: "<:EazyBotsCalender:838384636140453898>",
-  discord: {
-    js: "<:discordjs:884175977180196874>",
-  },
-  discordjs: "<:discordjs:884175977180196874>",
-  channel: "<:EazyBotsChannel:884170591685468160>",
-  members: "<:EazyBotsMembers:838385011052380180>",
-  verifiedBot: "<:EazyBotsBot:838387377013522442>",
-  developer: "<:EazyBotsDeveloper:838389685289484289>",
-  admin: "<a:EazyBotsAdmin:838393676639371274>",
-  pin: "<a:EazyBotsPin:838393671698481202>",
-  status: {
-    DND: "<:dnd:923577402184040458>",
-    OFFLINE: "",
-    IDLE: "<:idle:923577401865302027>",
-    ONLINE: "<:online:923577402137927760>",
-  },
-};
-
-export default emojis;
