+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "@eazyautodelete/eazyautodelete-core",
-  "version": "1.3.1",
+  "version": "1.3.5",
   "description": "🧰 Core Package used by the EazyAutodelete Discord Bot",
-  "main": "build/src/index.js",
+  "main": "build/index.js",
   "scripts": {
     "prepare": "husky install",
     "format": "npx prettier --write .",
@@ -23,7 +23,7 @@
     "type": "git",
     "url": "git+https://github.com/eazyautodelete/eazyautodelete-core.git"
   },
-  "typings": "build/src/index.d.ts",
+  "typings": "build/index.d.ts",
   "author": {
     "email": "dev@qreepex.xyz",
     "name": "qreepex",
@@ -35,6 +35,7 @@
   },
   "homepage": "https://github.com/eazyautodelete/eazyautodelete-core#readme",
   "devDependencies": {
+    "@types/json5": "^2.2.0",
     "@types/node": "^17.0.23",
     "@typescript-eslint/eslint-plugin": "^5.17.0",
     "@typescript-eslint/parser": "^5.17.0",
@@ -46,12 +47,15 @@
     "typescript": "^4.6.3"
   },
   "dependencies": {
-    "@eazyautodelete/eazyautodelete-lang": "^1.2.0",
-    "@types/mongoose": "^5.11.97",
+    "@discordjs/rest": "^0.5.0",
+    "@eazyautodelete/bot-utils": "^1.0.2",
+    "@eazyautodelete/eazyautodelete-db-client": "^1.2.3",
+    "@eazyautodelete/eazyautodelete-lang": "^1.2.2",
     "axios": "^0.24.0",
     "chalk": "^4.1.2",
     "discord-api-types": "^0.30.0",
-    "discord.js": "^13.8.0"
+    "discord.js": "^13.8.0",
+    "json5": "^2.2.1"
   },
   "publishConfig": {
     "registry": "https://npm.pkg.github.com/eazyautodelete"
