--- utils/colorToInt.ts
@@ -1,5 +1,0 @@
-import { ColorResolvable } from "discord.js";
-
-export default function (hex: ColorResolvable): number {
-  return parseInt(hex as string, 16);
-}
